Switch to parameterized tests
